+++ packages/next/src/server/dev/turbopack/manifest-loader.ts
@@ -7,7 +7,7 @@ import type { AppBuildManifest } from '../../../build/webpack/plugins/app-build-
 import type { PagesManifest } from '../../../build/webpack/plugins/pages-manifest-plugin'
 import { pathToRegexp } from 'next/dist/compiled/path-to-regexp'
 import type { ActionManifest } from '../../../build/webpack/plugins/flight-client-entry-plugin'
-import { generateRandomActionKeyRaw } from '../../app-render/action-encryption-utils'
+import { generateEncryptionKeyBase64 } from '../../app-render/encryption-utils'
 import type { NextFontManifest } from '../../../build/webpack/plugins/next-font-manifest-plugin'
 import type { LoadableManifest } from '../../load-components'
 import {
@@ -123,7 +123,7 @@ export class TurbopackManifestLoader {
     const manifest: ActionManifest = {
       node: {},
       edge: {},
-      encryptionKey: await generateRandomActionKeyRaw(true),
+      encryptionKey: await generateEncryptionKeyBase64(true),
     }
 
     function mergeActionIds(
